Just the other day I discovered [`'$handle_undefined_function'(FnName, Args)`](https://erlang.org/doc/man/error_handler.html#undefined_function-3) and it seems perfect to abuse! So I tried it and it was possible to cut the time to half again, never would have expected that. I extended the table in the previous PRs and it is now:

Execution time  | Code
----------------|--------------------------
67s             | Baseline (1.2.0)
32s             | PR #18
9s              | PRs #21 and #20
4s              | This PR

The biggest of the mocked modules in this eunit test had approx 4500 exported functions, and making the mocking module for it previously took 3.3s. It is now down to 0.1s.

The `'$handle_undefined_function'/2` seems to have been introduced roughly in R16, according to my archæological excavations.

As usual, I suppose I should test it a bit more thorougly on some largish projects first, so I'm creating a draft request to indicate that. But comments are of course welcome any time.