@tomas-abrahamsson 2021-10-04
I have run this PR plus the #26 (aa791d1111) on top in 4 large projects, and it looks good.